+++ config/filament-trumbowyg.php
@@ -11,4 +11,5 @@ return [
     'url_protocol' => false,
     'minimal_links' => false,
     'link_targets' => ['_self', '_blank'],
+    'image_width_modal_edit' => false,
 ];